Meet the React component
In contrast to traditional JavaScript functions, components are
defined in PascalCase and have to start with a capital letter otherwise
they are treated as simple HTML elements.

  Since we do not want to re-define a variable within a function every
time this function runs, we could define this variable outside of the
component as well. In this case, the title does not depend on any
information that's within the function component (e.g. parameters coming
from the function signature), hence it's okay to move it outside.
Therefore it will be defined only once and not every time the function
is called.

  As a rule of thumb:
If a variable does not need anything from within the function component
body (e.g. parameters), then define it outside of the component which
avoids re-defining it every function call.