Create component with React.createElement
JSX is a syntax extension to JavaScript. In the past, JavaScript files
which made use of JSX had to use the .jsx instead of the .js
extension. However, these days the underlying build tools
(read: compiler/bundler) can be configured to acknowledge JSX in a .js
file. If they are configured this way, they will transpile JSX to
JavaScript. Tools like Vite embrace the .jsx extension though, because
it makes it more explicit for developers.