Lists in React
Usually using an index should be avoided though, because it comes with
the same rendering performance issues from above. In addition, it can
cause actual bugs in the UI⁵¹ whenever the order of items got changed
(e.g. re-ordering, appending or removing items). However, as last
resort, if the list does not change its order in any way, using the
index is fine.