Callback handler
The concept of the callback handler in a nutshell: We pass a function
from a parent component (App) to a child component (Search) via props;
we call this function in the child component (Search), but have the
actual implementation of the called function in the parent component
(App). In other words, when an (event) handler is passed as props from
a parent component to its child component, it becomes a callback
handler.

React props are always passed down the component tree and
therefore functions that are passed down as callback handlers in props
can be used to communicate up the component tree.