Controled components
Earlier the HTML element did its own thing, but now we are in control
of it by feeding React's state into it. Now, while the input field
became explicitly a controlled element, the Search component became
implicitly a controlled component. As a React beginner, using
controlled components is important, because you want to enforce a
predictable behavior