Side effects
The unique purpose of a React component is to render UI. Not running
side effects which the tasks of the side effet hook.

React’s useEffect Hook takes two arguments: The first argument is a
function that runs our side effect. In our case, the side-effect stores
searchTerm into the browser’s local storage. The second argument is a
dependency array of variables. If one of these variables changes, the
function for the side-effect is called. In our case, the function is
called every time the searchTerm changes. In addition, it’s also called
initially when the component renders for the first time.

Leaving out the second argument (the dependency array) would make the
function for the side-effect run on every render
(initial render and update renders) of the component.

If the dependency array of React’s useEffect is an empty array, the
function for the side-effect is only called once when the component
renders for the first time.

After all, the hook lets us opt into React’s component lifecycle.
It can be triggered when the component is first mounted, but also if
one of its values (state, props, derived values from state/props) is
updated.

In conclusion, using React useEffect Hook instead of managing the
side-effect in the (event) handler has made the application more robust.
Whenever and wherever the searchTerm state is updated via setSearchTerm,
the browser’s local storage will always be in sync with it.