Reusable component
It's always a trade-off between generalization and specialization of
components. In this case, we turned a highly specialized component into
a generalized component. While a generalized component has a better
chance of getting reused in the application, a specialized component
would implement business logic for one specific use case and therefore
isn't reusable at all.