Inline handlers
Now we need to make each list item removable by prividing it with a
dismiss button.

To be effective the item must send its ID to the filtering parent
by calling the parent's callback handlers with the required arguments
without triggering it out of command. It is achieved both ways:
	{ callbackHandler.bind(null, argument) } or simpler,
	{ () => callbackHandler(argument) }

It's okay to use inline handlers if they do not obscure critical
implementation details. If inline handlers need to use a block body,
because there are more than one line of code executed, it's about time
to extract them as normal event handlers.