Advanced state: useReducer
Using useReducer over useState makes sense as soon as multiple stateful
values are dependent on each other or related to one domain. If a
reducer case is not covered, throw an error to remind yourself that the
implementation isn't covered. Let's bring the concept of a reducer to a
minimal version by handling more than one state transition. If there is
only one state transition, a reducer wouldn't make sense.