Impossible state
The impossible state happens when an error occurs for the asynchronous
data. The state for the error is set, but the state for the loading
indicator isn't revoked. In the UI, this would lead to an infinite
loading indicator and an error message, though it may be better to show
the error message only and hide the loading indicator. Impossible
states are not easy to spot, which makes them infamous for causing bugs
in the UI. You could go on and try yourself to fix this bug.