Memoized handlers with useCallback
By moving the data fetching function outside the React's `useEffect`
hook, it becomes reusable for other parts of the application. We won't
use it just yet, but it is a good use   case to understand the memoized
functions in React. Now the `useEffect` hook runs implicitly  when the
searchTerm changes, because the `handleFetchStories` is re-defined each
time the searchTerm changes. Since the `useEffect` hook depends on the
`handleFetchStories`, the side- effect for data fetching runs again.

If we didn't create a memoized function with React's `useCallback` Hook,
a new `handleFetchStories` function would be created each time the `App`
component re-renders, and would be executed in the `useEffect` hook to
fetch data. The fetched data then, stored as state in the component,
would trigger the component re-renders which in turn would re-create a
new `handleFetchStories` function. The side-effect would run again and
the `App` would be stuck in an endless loop.