Explicit data feching
Before the `searchTerm` was used for two cases: updating the input
field's state and activating the side effect for fetching data. Now it's
only used for the former. A second state called url got introduced to
trigger the side-effect that fetches the data which only happens when a
user clicks the confirmation button.

Notice: a stateful value (e.g. url) can be INITIALIZE with another
stateful value.

Question: Why is useState instead of useStorageState used for the url
state management?
Answer: We do not want to remember the url in the browser's local
storage, because it's already derived from a static string
(here: API_ENDPOINT) and the searchTerm which already comes from the
browser's local storage.

Question: Why is there no check for an empty searchTerm in the
handleFetchStories function anymore?
Answer: Preventing a server-side search happens in the new button,
because it gets disabled whenever there is no searchTerm.