Third party libraries
We previously introduced the native fetch API
(which the browser provides) to perform requests to the Hacker News API.
However, not all browsers support this, especially the older ones. Also,
once you start testing your application in a headless browser
environment, issues can arise with the fetch API, because the actual
browser is not there.

There are a couple of ways to make fetch work in older browsers
(polyfills) and in tests (isomorphic fetch), but these concepts are a bit
off-task for the purpose of this learning experience.

We can see with these examples that axios is a powerful library for
performing requests to remote APIs. I recommend it over the native fetch
API when requests become complex, working with older browsers, or for
testing.