Forms: submit with <CR>
Forms aren't much different in React than in plain HTML. When we have
input fields and a button to submit data from them, we can give our HTML
more structure by wrapping it into a form element with a onSubmit
attribute. The button that executes the submission therefore needs the
"submit" type to refer the process to the form element's handler.
After all, it makes it more accessible for keyboard users as well.