Strict mode
So React's Strict Mode runs React's useEffect Hooks twice for the initial
render. Because this results in fetching the same data twice, this is not
a problem for us. The operation is called idempotent, which means that
the result of a successfully performed request is independent of the
number of times it is executed. After all, it's only a performance
problem, because there are two network requests, but it doesn't result in
a buggy behavior of the application. In addition to all of this
uncertainty, the Strict Mode is only applied for the development
environment, so whenever this application gets build for production, the
Strict Mode gets removed automatically.