Memoize expensive computation result
`React.useMemo` hook:
If all arguments are passed to a function, it's acceptable to have it
outside the component, because it does not have any further dependency
needed from within the component. This prevents creating the function on
every render, so the useCallback hook becomes unnecessary.
For every time someone types in the SearchForm, the computation shouldn't
run again. It only runs if the dependency array, here stories, has
changed. After all, this should only be used for cost expensive
computations which could lead to a delay of a (re-)rendering of a
component. Now, after we went through these scenarios for useMemo,
useCallback, and memo, remember that these shouldn't necessarily be used
by default. Apply these performance optimizations only if you run into
performance bottlenecks. Most of the time this shouldn't happen, because
React's rendering mechanism is pretty efficient by default. Sometimes the
check for utilities like memo can be more expensive than the re-rendering
itself.