SVGs in React TS
The Vite plugin makes using SVGs straightforward, with not much extra configuration needed. This is different if you create a React project from scratch with build tools like Webpack, because you have to take care of it yourself.

It's enough to
(1) import and call it in your vite.config.ts file
(2) in the using module import the SVG appending the extension "?react"
(3) if using typescript ensure proper type reference at the top of all
importing modules with:
/// <reference types="vite-plugin-svgr/client" />